+++ src/leds.c
@@ -27,12 +27,12 @@ void leds_shift_init(void) {
 	shift a pin state out.  pins_to_shift is how many bits to shift
 	buffering with zero instead of the new_state bit values
 */
-void leds_change_state(uint32_t new_state, uint8_t pins_to_shift) {
+void leds_change_state(uint64_t new_state, uint8_t pins_to_shift) {
 	LEDS_PORT &= ~_BV(LEDS_RCLK);
 
 	for (uint8_t i = 0; i < LEDS_SHIFT_REG_COUNT * 8; i++) {
 		LEDS_PORT &= ~_BV(LEDS_SRCLK);
-		if (pins_to_shift && new_state & ((uint32_t)1 << i)) {
+		if (pins_to_shift && new_state & ((uint64_t)1 << i)) {
 			LEDS_PORT |= _BV(LEDS_SER);
 			pins_to_shift--;
 		}
